Tests/LibWeb/Text/expected/DOM/Offset-of-empty-inline-element.txt

  Top: 0
Left: 0
